ex00 and ex01 fixed

+++ ex01/RPN.cpp
@@ -10,31 +10,31 @@ RPN::~RPN() {}
 RPN::RPN(const RPN &other) { (void)other; }
 RPN& RPN::operator=(const RPN &other) { (void)other; return (*this); }
 
-int RPN::addition(int a, int b) { return a + b; }
-int RPN::subtraction(int a, int b) { return a - b; }
-int RPN::multiplication(int a, int b) { return a * b; }
-int RPN::division(int a, int b) {
+float RPN::addition(float a, float b) { return a + b; }
+float RPN::subtraction(float a, float b) { return a - b; }
+float RPN::multiplication(float a, float b) { return a * b; }
+float RPN::division(float a, float b) {
 	if (b == 0)
 		throw std::runtime_error("Cannot divide by 0");
 	return a / b; 
 }
 
-int RPN::calculate(std::string input) {
+float RPN::calculate(std::string input) {
 	if (input.empty())
 		throw std::runtime_error("Empty input");
 	
-	std::stack<int> stack;
+	std::stack<float> stack;
 
 	for (std::size_t i = 0; i < input.size(); ++i) {
 		if (std::isdigit(static_cast<unsigned char>(input[i]))) {
-			stack.push(static_cast<int>(input[i] - '0'));
+			stack.push(static_cast<float>(input[i] - '0'));
 		}
 		else {
 			if (stack.size() < 2)
 				throw std::runtime_error("Not enough digits");
-			int b = stack.top();
+			float b = stack.top();
 			stack.pop();
-			int a = stack.top();
+			float a = stack.top();
 			stack.pop();
 			switch (input[i]) {
 				case '+': stack.push(addition(a, b)); break;
